Update lmo-tippcalcgraph.php
PHP Warning:  A non-numeric value encountered in ../lmo/addon/tipp/lmo-tippcalcgraph.php on line 120
Bei eingeloggtem User und Aufruf der Fieberkurven

+++ lmo/addon/tipp/lmo-tippcalcgraph.php
@@ -117,7 +117,7 @@ if (!file_exists($auswertfile)) {
     $dum=explode('|',$tippdaten[$i-1]);
     $op1=$dum[0];
     // Nick
-    $op3=substr($dum[1],2)-1;
+    $op3=(int)substr($dum[1],2)-1;
     // Spieltagsnummer
     $op4=substr($dum[1],0,2);
     // TP
